requirements.txt
# Core
Django
django-allauth
psycopg[binary]
django-environ
dj-database-url
whitenoise
django-htmx
django-debug-toolbar
gunicorn
python-dotenv

# Multitenancy
django-tenants

# Workflow / state
django-fsm-2
django-fsm-log

# Dev & test
pytest
pytest-django
pytest-cov
model-bakery
black
isort
flake8

pillow # Image processing
djlint # Code formatting for Django templates